feat: add experimental offcpu capabilities  to eBPF profiler
Tracks off-cpu time and resolve symbols found in the stack traces.
The ultimate goal is to provide flamegraphs to help visualizing where
the profiled programs are spending time outside of the cpu.

For now, this is an experimental feature that must be enabled with
--features ebpf; I will see if I continue this rabbit hole that I know
basically nothing about.

This PR includes:

- Error detection and reporting for stack trace failures
- Memory map caching with better refresh logic
- Added robust symbolication with multiple resolution methods

1. Interpreter/JIT languages (Python, Java, Node.js) will still have
   limited stack trace support due to their unique memory layouts
2. Kernel stack symbolication requires additional setup (not covered
   here)
3. Some stack frames may still fail to symbolicate if debug info is
   insufficient